test: StateHolder 유닛 테스트 추가
- CardListStateHolderTest: 카드 목록 상태 관리 로직 테스트
- RegisterCardStateHolderTest: 카드 등록 상태 관리 로직 테스트

refactor: CardUiModel의 bankColor 타입 Int에서 Long으로 변경

- CardUiModel의 bankColor 프로퍼티 타입을 Int에서 Long으로 변경하여 ARGB 값을 직접 저장하도록 수정했습니다.
- 이에 따라 Card 도메인 모델을 CardUiModel로 변환하는 로직 및 프리뷰 코드도 수정했습니다.

docs: 기능 구현 목록 업데이트

- (선택사항) 카드사를 선택할 때 적절한 카드사 아이콘을 노출한다. 항목 완료 처리

+++ app/src/main/java/woowacourse/payments/ui/model/CardUiModel.kt
@@ -12,7 +12,7 @@ data class CardUiModel(
     val expirationDate: String,
     val cardHolderName: String,
     val bankName: String,
-    val bankColor: Int,
+    val bankColor: Long,
 ) : Parcelable
 
 fun Card.toUiModel(): CardUiModel {
@@ -24,7 +24,11 @@ fun Card.toUiModel(): CardUiModel {
         expirationDate = formattedExpirationDate,
         cardHolderName = this.cardHolderName.cardHolderName,
         bankName = this.bank.name,
-        bankColor = this.bank.color().toArgb(),
+        bankColor =
+            this.bank
+                .color()
+                .toArgb()
+                .toLong(),
     )
 }
